requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cog==0.7.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.3
ffmpeg==1.4
fonttools==4.39.3
fqdn==1.5.1
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
ipykernel==6.23.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
munkres==1.1.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
opencv-contrib-python==4.7.0.72
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
pooch==1.4.0
progress==1.6
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
PyWavelets==1.4.1
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.29.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Send2Trash==1.8.2
setuptools==66.0.0
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.26.1
structlog==22.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
vidstab==1.7.4
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
wheel==0.38.4
widgetsnbextension==4.0.7
win-inet-pton==1.1.0
zipp==3.15.0
